// src/shared/api/generated/schemas/adminWorkspacesListResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Smetchik API
 * Smetchik service API documentation
 * OpenAPI spec version: 1.0.0
 */
import type { Workspace } from './workspace';

export interface AdminWorkspacesListResponse {
  workspaces?: Workspace[];
  total?: number;
  limit?: number;
  offset?: number;
}