// src/shared/api/generated/schemas/workspacesListResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Smetchik API
 * Smetchik service API documentation
 * OpenAPI spec version: 1.0.0
 */
import type { Workspace } from './workspace';

export interface WorkspacesListResponse {
  workspaces?: Workspace[];
  /** Total number of workspaces */
  total?: number;
  limit?: number;
  offset?: number;
}